+++ lib/controllers/auth.js
@@ -71,8 +71,6 @@ app.get("/facebook/token", passport.authenticate("facebook", { session: false })
         if (err) {
             return next(err);
         }
-
-        console.log(errors.ERRORS.USER_NOT_REGISTERED);
         if (!user) {
             return next(errors.ERRORS.USER_NOT_REGISTERED);
         }
